Add simply learning capabilities to allow predictions

+++ src/main/kotlin/org/fullsack101/bandit/predictor/ModelReward.kt
@@ -1,4 +1,6 @@
 package org.fullsack101.bandit.predictor
 
+import java.math.BigDecimal
+
 @JvmInline
-value class ModelReward(val reward: Float)
+value class ModelReward(val reward: BigDecimal)
